fix: move and tweak packages_to_string
