+++ crates/flox/src/utils/mod.rs
@@ -17,6 +17,7 @@ use once_cell::sync::Lazy;
 pub mod colors;
 mod completion;
 pub mod dialog;
+pub mod display;
 pub mod init;
 pub mod installables;
 pub mod logger;
